rename package 'infolink.datasetMatcher' to 'resolve', #45
