+++ src/main/java/io/github/infolis/model/Execution.java
@@ -7,6 +7,7 @@ import io.github.infolis.algorithm.SearchTermPosition;
 import io.github.infolis.algorithm.TextExtractor;
 import io.github.infolis.datastore.DataStoreClient;
 import io.github.infolis.datastore.FileResolver;
+import io.github.infolis.resolve.QueryService;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
@@ -23,7 +24,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import io.github.infolis.infolink.datasetMatcher.QueryService;
+
 import java.lang.reflect.Field;
 
 /**
