+++ src/main/java/io/github/infolis/resolve/DaraHTMLQueryService.java
@@ -1,5 +1,5 @@
 
-package io.github.infolis.infolink.datasetMatcher;
+package io.github.infolis.resolve;
 
 /**
  *
