+++ src/main/java/io/github/infolis/resolve/DaraWebMatcher.java
@@ -1,4 +1,4 @@
-package io.github.infolis.infolink.datasetMatcher;
+package io.github.infolis.resolve;
 
 import io.github.infolis.util.RegexUtils;
 import io.github.infolis.util.URLParamEncoder;
