+++ src/main/java/io/github/infolis/resolve/SolrQueryService.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package io.github.infolis.infolink.datasetMatcher;
+package io.github.infolis.resolve;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
