+++ src/test/java/io/github/infolis/algorithm/FederatedSearcherTest.java
@@ -1,12 +1,13 @@
 package io.github.infolis.algorithm;
 
 import io.github.infolis.InfolisBaseTest;
-import io.github.infolis.infolink.datasetMatcher.DaraHTMLQueryService;
-import io.github.infolis.infolink.datasetMatcher.DaraSolrQueryService;
-import io.github.infolis.infolink.datasetMatcher.QueryService;
 import io.github.infolis.model.Execution;
 import io.github.infolis.model.SearchQuery;
 import io.github.infolis.model.entity.SearchResult;
+import io.github.infolis.resolve.DaraHTMLQueryService;
+import io.github.infolis.resolve.DaraSolrQueryService;
+import io.github.infolis.resolve.QueryService;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
