+++ src/test/java/io/github/infolis/algorithm/SolrQueryServiceTest.java
@@ -1,11 +1,12 @@
 package io.github.infolis.algorithm;
 
 import io.github.infolis.InfolisBaseTest;
-import io.github.infolis.infolink.datasetMatcher.QueryService;
-import io.github.infolis.infolink.datasetMatcher.SolrQueryService;
 import io.github.infolis.model.Execution;
 import io.github.infolis.model.SearchQuery;
 import io.github.infolis.model.entity.SearchResult;
+import io.github.infolis.resolve.QueryService;
+import io.github.infolis.resolve.SolrQueryService;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
