+++ src/test/java/io/github/infolis/resolve/DaraSolrMatcherTest.java
@@ -1,4 +1,4 @@
-package io.github.infolis.infolink.datasetMatcher;
+package io.github.infolis.resolve;
 
 import static org.junit.Assert.assertEquals;
 
@@ -14,6 +14,8 @@ import org.junit.Assume;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import io.github.infolis.resolve.DaraSolrMatcher;
 
 public class DaraSolrMatcherTest {
